Bluetooth: Inform Power Manager about HID Host Connection.
This patch informs Power Manager about HID host connection complete
status which in-turn starts the inactivity timer of 30 seconds for
sending the sniff command on timeout if the ACL mode is still active
on timeout. Without this patch, if remote device is not initiating
sniff on HID ACL, host was also not initiating sniff, resulting in
unnecessary power consumption on ideal HID Link.

Change-Id: Icbe34e855112800b07f98f40dd7c6ae5e91c4cf6
CRs-Fixed: 531573